Day 3: Rucksack Reorganization (#4)
The task of the third puzzle was to help the elves organize their
packbags. The first part identified misplaces items, while the second
was about finding shared items across multiple packbags.

The implementation has not been optimized for speed or efficiency. There
are multiple loops that can be shortened, but given the input size this
felt like an over-optimization.